// apps/frontend/src/features/recipes/recipes.types.ts
export type RecipeFormState = {
  recipeName: string
  author: string
  url: string
  cameraSensor: string
  film: string
  grain: string
  colorChromeEffect: string
  colorChromeFxBlue: string
  whiteBalance: string
  dynamicRange: string
  highlight: string
  shadow: string
  color: string
  sharpness: string
  nr: string
  clarity: string
  expCompensation: string
  iso: string
  notes: string
}